pyproject.toml
[project]
name = "rain-radar-telegram-bot"
version = "0.1.0"
description = "AWS Lambda worker that polls Singapore radar images and pushes rain alerts to Telegram"
requires-python = ">=3.11"
dependencies = [
  "boto3>=1.34.0",
  "pypng>=0.20220715.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
